+++ core-js/blocks-filters/responsive-settings/settings/container/grid/properties.ts
@@ -82,6 +82,19 @@ const getAlignContentValues = () => [
   { label: __('Stretch', 'webentor'), value: 'content-stretch' },
 ];
 
+const getOrderValues = () => [
+  { label: __('None selected', 'webentor'), value: '' },
+  { label: __('Order First', 'webentor'), value: 'order-first' },
+  { label: __('Order Last', 'webentor'), value: 'order-last' },
+  { label: __('Order None', 'webentor'), value: 'order-none' },
+  { label: __('Order 0', 'webentor'), value: 'order-0' },
+  { label: __('Order 1', 'webentor'), value: 'order-1' },
+  { label: __('Order 2', 'webentor'), value: 'order-2' },
+  { label: __('Order 3', 'webentor'), value: 'order-3' },
+  { label: __('Order 4', 'webentor'), value: 'order-4' },
+  { label: __('Order 5', 'webentor'), value: 'order-5' },
+];
+
 const getGridColumnSpanValues = (twTheme: WebentorConfig['theme']) => {
   const values = Object.keys(twTheme?.gridColumn).map((key) => ({
     label: key,
@@ -165,5 +178,10 @@ export const getGridItemProperties = (twTheme: WebentorConfig['theme']) => [
     label: __('Grid Row Span', 'webentor'),
     name: 'grid-row-span',
     values: getGridRowSpanValues(twTheme),
+  },
+  {
+    label: __('Order', 'webentor'),
+    name: 'order',
+    values: getOrderValues(),
   },
 ];